+++ CMakeModules/Findmc_rtc_3rd_party_geos.cmake
@@ -28,10 +28,10 @@ if(NOT TARGET GEOS::geos)
     )
     if(GEOS_CONFIG)
       # Get GEOS_INSTALL_PREFIX from geos-config
-      exec_program(
-        ${GEOS_CONFIG} ARGS
-        --prefix
+      execute_process(
+        COMMAND ${GEOS_CONFIG} --prefix
         OUTPUT_VARIABLE GEOS_INSTALL_PREFIX
+        OUTPUT_STRIP_TRAILING_WHITESPACE
       )
       find_library(GEOS_LIBRARY NAME geos HINTS ${GEOS_INSTALL_PREFIX})
       if(NOT GEOS_LIBRARY)