+++ src/mc_rtc/Logger.cpp
@@ -213,7 +213,8 @@ void Logger::start(const std::string & ctl_name, double timestep, bool resume, d
   {
     if(resume)
     {
-      // Repeat the added key events
+      // Re-create key events based on the current set of entries
+      log_events_.clear();
       for(const auto & e : log_entries_) { log_events_.push_back(KeyAddedEvent{e.type, e.key}); }
     }
     else { impl_->log_iter_ = start_t; }
@@ -297,12 +298,14 @@ void Logger::log()
       }
       else if constexpr(std::is_same_v<T, StartEvent>)
       {
-        builder.start_array(5);
+        builder.start_array(7);
         builder.write(static_cast<uint8_t>(3));
         builder.write(meta_.timestep);
         builder.write(meta_.main_robot);
         builder.write(meta_.main_robot_module);
         builder.write(meta_.init);
+        builder.write(meta_.init_q);
+        builder.write(meta_.calibs);
         builder.finish_array();
       }
       else { static_assert(!std::is_same_v<T, T>, "non-exhaustive visitor"); }