+++ src/mc_tasks/VectorOrientationTask.cpp
@@ -211,4 +211,4 @@ static auto registered = mc_tasks::MetaTaskLoader::register_load_function(
       t->load(solver, config);
       return t;
     });
-}
+} // namespace