+++ src/mc_tasks/lipm_stabilizer/StabilizerTask.cpp
@@ -274,11 +274,18 @@ void StabilizerTask::enable()
   configure(lastConfig_);
   zmpcc_.enabled(true);
   enabled_ = true;
+  wasEnabled_ = true;
 }
 
 void StabilizerTask::disable()
 {
   mc_rtc::log::info("[StabilizerTask] disabled");
+  disable_();
+  wasEnabled_ = false;
+}
+
+void StabilizerTask::disable_()
+{
   // Save current configuration to be reused when re-enabling
   lastConfig_ = c_;
   disableConfig_ = c_;
@@ -523,6 +530,25 @@ void StabilizerTask::checkInTheAir()
   {
     inTheAir_ = inTheAir_ && footT.second->measuredWrench().force().z() < c_.safetyThresholds.MIN_DS_PRESSURE;
   }
+
+  if(!wasInTheAir_ && inTheAir_)
+  {
+    wasInTheAir_ = true;
+    if(enabled_)
+    {
+      mc_rtc::log::warning("[{}] Robot is in the air, disabling stabilizer", name());
+      disable_();
+    }
+  }
+  else if(!inTheAir_ && wasInTheAir_)
+  {
+    wasInTheAir_ = false;
+    if(wasEnabled_)
+    {
+      mc_rtc::log::warning("[{}] Robot is no longer in the air, re-enabling stabilizer", name());
+      enable();
+    }
+  }
 }
 
 void StabilizerTask::computeLeftFootRatio()
@@ -732,7 +758,7 @@ void StabilizerTask::run()
       footTasks[ContactState::Left]->setZeroTargetWrench();
     }
   }
-  distribZMP_ = mc_rbdyn::zmp(distribWrench_, zmpFrame_);
+  if(!mc_rbdyn::zmp(distribZMP_, distribWrench_, zmpFrame_)) { return; }
 
   updateCoMTaskZMPCC();
   updateFootForceDifferenceControl();
@@ -1579,4 +1605,4 @@ static auto registered = mc_tasks::MetaTaskLoader::register_load_function(
       t->reset();
       return t;
     });
-}
+} // namespace